+++ packages/plugins/did-web-plugin/CHANGELOG.md
@@ -1,5 +1,19 @@
 # did-web-plugin
 
+## 1.0.74
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/core@9.3.37
+
+## 1.0.73
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/core@9.3.36
+
 ## 1.0.72
 
 ### Patch Changes